BATCH-1869: Removed StringDelimitedLineTokenizer in favor of updating existing class
