+++ spring-batch-samples/src/main/resources/jobs/headerFooterSample.xml
@@ -33,7 +33,7 @@
 			<bean class="org.springframework.batch.item.file.mapping.DefaultLineMapper">
 				<property name="lineTokenizer">
 					<bean
-						class="org.springframework.batch.item.file.transform.StringDelimitedLineTokenizer">
+						class="org.springframework.batch.item.file.transform.DelimitedLineTokenizer">
 						<property name="delimiter" value="," />
 					</bean>
 				</property>
