+++ spring-batch-samples/src/main/resources/jobs/iosample/multiLine.xml
@@ -28,7 +28,7 @@
 				<property name="lineMapper">
 					<bean class="org.springframework.batch.item.file.mapping.DefaultLineMapper">
 						<property name="lineTokenizer">
-							<bean class="org.springframework.batch.item.file.transform.StringDelimitedLineTokenizer" />
+							<bean class="org.springframework.batch.item.file.transform.DelimitedLineTokenizer" />
 						</property>
 						<property name="fieldSetMapper">
 							<bean class="org.springframework.batch.item.file.mapping.PassThroughFieldSetMapper" />
