+++ spring-batch-samples/src/main/resources/jobs/iosample/multiResource.xml
@@ -18,7 +18,7 @@
 						class="org.springframework.batch.item.file.mapping.DefaultLineMapper">
 						<property name="lineTokenizer">
 							<bean
-								class="org.springframework.batch.item.file.transform.StringDelimitedLineTokenizer">
+								class="org.springframework.batch.item.file.transform.DelimitedLineTokenizer">
 								<property name="delimiter" value="," />
 								<property name="names" value="name,credit" />
 							</bean>
