+++ spring-batch-samples/src/main/resources/jobs/skipSampleJob.xml
@@ -77,7 +77,7 @@
 		<property name="lineMapper">
 			<bean class="org.springframework.batch.item.file.mapping.DefaultLineMapper">
 				<property name="lineTokenizer">
-					<bean class="org.springframework.batch.item.file.transform.StringDelimitedLineTokenizer">
+					<bean class="org.springframework.batch.item.file.transform.DelimitedLineTokenizer">
 						<property name="names" value="ISIN, Quantity, Price, Customer" />
 					</bean>
 				</property>
